# dpi-frontend/src/app/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Swagger Api documentation",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "Swagger Api Documentation",
    "contact" : {
      "name" : "SysInfoMed",
      "url" : "https://www.sysinfomed.com",
      "email" : "support@sysinfomed.com"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/",
...
  "paths" : {
...
    "/api/v1/address/{addressId}" : {
      "get" : {
        "tags" : [ "Address" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "addressId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AddressDto"
            }
          }
        }
      }
    },
...
    "/api/v1/allergies/{allergyId}" : {
      "get" : {
        "tags" : [ "Allergies" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "allergyId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AllergyDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Allergies" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "allergyId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/allergy-types/{allergyTypeId}" : {
      "get" : {
        "tags" : [ "Allergy types" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "allergyTypeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AllergyTypeDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Allergy types" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "allergyTypeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/appointments/{id}" : {
      "get" : {
        "tags" : [ "Appointment" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/AppointmentDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Appointment" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api/v1/bookings/{id}" : {
      "get" : {
        "tags" : [ "Booking" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BookingDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Booking" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api/v1/constant-types/{constantTypeId}" : {
      "get" : {
        "tags" : [ "Constants types" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "constantTypeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ConstantTypeDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Constants types" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "constantTypeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/constants/{constantId}" : {
      "get" : {
        "tags" : [ "Constants" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "constantId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ConstantDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Constants" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "constantId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/consultations/{consultationId}" : {
      "get" : {
        "tags" : [ "Consultation" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "consultationId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ConsultationDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Consultation" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "consultationId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/api/v1/contact-person/{contactPersonId}" : {
      "get" : {
        "tags" : [ "Contact-Person" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "contactPersonId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ContactPersonDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Contact-Person" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "contactPersonId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/hospitals/{hospitalId}" : {
      "get" : {
        "tags" : [ "hospital" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "hospitalId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/HospitalDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "hospital" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "hospitalId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/medical-background-types/{medicalBackgroundTypeId}" : {
      "get" : {
        "tags" : [ "Medical background types" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "medicalBackgroundTypeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MedicalBackgroundTypeDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Medical background types" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "medicalBackgroundTypeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/medical-backgrounds/{medicalBackgroundId}" : {
      "get" : {
        "tags" : [ "Medical backgrounds" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "medicalBackgroundId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MedicalBackgroundDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Medical backgrounds" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "medicalBackgroundId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/medical-case-types/{medicalCaseTypeId}" : {
      "get" : {
        "tags" : [ "medical cases types" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "medicalCaseTypeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MedicalCaseTypeDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "medical cases types" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "medicalCaseTypeId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/medical-cases/{caseId}" : {
      "get" : {
        "tags" : [ "Medical cases" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MedicalCaseDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Medical cases" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "caseId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/medical-services/{medicalServiceId}" : {
      "get" : {
        "tags" : [ "medical service" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "medicalServiceId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MedicalServiceDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "medical service" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "medicalServiceId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/api/v1/medications/" : {
      "get" : {
        "tags" : [ "Medication" ],
        "operationId" : "findAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/MedicationDto"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Medication" ],
        "operationId" : "save",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/MedicationDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        }
      }
    },
    "/api/v1/medications/{medicationId}" : {
      "get" : {
        "tags" : [ "Medication" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "medicationId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/MedicationDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Medication" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "medicationId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/api/v1/preferences/{preferenceId}" : {
      "get" : {
        "tags" : [ "Preferences cases" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "preferenceId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PreferenceDto"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Preferences cases" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "preferenceId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "AddressDto" : {
      "type" : "object",
      "required" : [ "street" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "street" : {
          "type" : "string"
        },
        "houseNumber" : {
          "type" : "string"
        },
        "country" : {
          "type" : "string"
        },
        "postalCode" : {
          "type" : "string"
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "AllergyDto" : {
      "type" : "object",
      "required" : [ "allergy", "allergyTypeId" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "allergy" : {
          "type" : "string"
        },
        "remarks" : {
          "type" : "string"
        },
        "confidential" : {
          "type" : "boolean"
        },
        "allergyTypeId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "allergyTypeLabel" : {
          "type" : "string"
        },
        "date" : {
          "type" : "integer",
          "format" : "int64"
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "AllergyTypeDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "allergyName" : {
          "type" : "string"
        }
      }
    },
    "AppointmentDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "personnelId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "consultationId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "medicalCaseId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "bookingId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "startTime" : {
          "$ref" : "#/definitions/LocalTime"
        },
        "endTime" : {
          "$ref" : "#/definitions/LocalTime"
        }
      }
    },
...
    "BookingDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "serviceId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "bookingDate" : {
          "type" : "string",
          "format" : "date"
        }
      }
    },
...
    "ConstantDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "value" : {
          "type" : "number",
          "format" : "double"
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "constantTypeId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "constantTypeLabel" : {
          "type" : "string"
        },
        "constantTypeUnit" : {
          "type" : "string"
        },
        "date" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "ConstantTypeDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "constantName" : {
          "type" : "string"
        },
        "constantUnit" : {
          "type" : "string"
        }
      }
    },
    "ConsultationDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "anamnesis" : {
          "type" : "string"
        },
        "physicalExamination" : {
          "type" : "string"
        },
        "complementaryExamination" : {
          "type" : "string"
        },
        "differentialDiagnosis" : {
          "type" : "string"
        },
        "proposition" : {
          "type" : "string"
        },
        "furtherConsultation" : {
          "type" : "string"
        },
        "appointmentId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "STAND_BY", "OPEN", "IN_PROGRESS", "CLOSED" ]
        }
      }
    },
    "ContactPersonDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "firstname" : {
          "type" : "string"
        },
        "lastname" : {
          "type" : "string"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "date" : {
          "type" : "integer",
          "format" : "int64"
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "GeoLocation" : {
      "type" : "object",
      "properties" : {
        "latitude" : {
          "type" : "integer",
          "format" : "int64"
        },
        "longitude" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "HospitalDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "PRIVATE", "PUBLIC" ]
        },
        "logo" : {
          "type" : "string"
        },
        "dashboardPhoto" : {
          "type" : "string"
        },
        "directorName" : {
          "type" : "string"
        },
        "locality" : {
          "type" : "string"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "webSiteUrl" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "location" : {
          "$ref" : "#/definitions/GeoLocation"
        }
      }
    },
...
    "LocalTime" : {
      "type" : "object",
      "properties" : {
        "hour" : {
          "type" : "integer",
          "format" : "int32"
        },
        "minute" : {
          "type" : "integer",
          "format" : "int32"
        },
        "second" : {
          "type" : "integer",
          "format" : "int32"
        },
        "nano" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "MedicalBackgroundDto" : {
      "type" : "object",
      "required" : [ "medicalBackgroundTypeId", "patientId" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "background" : {
          "type" : "string"
        },
        "confidential" : {
          "type" : "boolean"
        },
        "period" : {
          "type" : "integer",
          "format" : "int32"
        },
        "treatment" : {
          "type" : "string"
        },
        "remarks" : {
          "type" : "string"
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "medicalBackgroundTypeId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "medicalBackgroundTypeLabel" : {
          "type" : "string"
        },
        "date" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "MedicalBackgroundTypeDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "backgroundName" : {
          "type" : "string"
        }
      }
    },
    "MedicalCaseDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "comment" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "OPEN" ]
        },
        "medicalCaseTypeId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "medicalCaseTypeTitle" : {
          "type" : "string"
        },
        "medicalCaseTypeDescription" : {
          "type" : "string"
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "date" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "MedicalCaseTypeDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "title" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "date" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "MedicalServiceDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string"
        },
        "code" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "director" : {
          "type" : "string"
        },
        "logo" : {
          "type" : "string"
        },
        "billable" : {
          "type" : "boolean"
        },
        "location" : {
          "$ref" : "#/definitions/GeoLocation"
        },
        "phoneNumber" : {
          "type" : "string"
        },
        "canHospitalize" : {
          "type" : "boolean"
        },
        "active" : {
          "type" : "boolean"
        },
        "hospitalId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "date" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "MedicationDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "code" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "price" : {
          "type" : "number"
        }
      }
    },
...
    "PreferenceDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string"
        },
        "patientId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "date" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
...
  }
}